article_service/requirements.txt
asgiref==3.8.1
black==24.4.2
cfgv==3.4.0
click==8.1.7
distlib==0.3.8
Django==5.0.7
djangorestframework==3.15.2
filelock==3.15.4
gunicorn==22.0.0
identify==2.6.0
mypy-extensions==1.0.0
nodeenv==1.9.1
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
pre-commit==3.8.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
PyYAML==6.0.1
sqlparse==0.5.1
virtualenv==20.26.3
